Add aliases for some other non-wikipedia projects
The Kiwix folks requests wikivoyage aliases, and we might as well add some
more projects to the list while we are at it. We are using the list of
Wikipedias for languages, which will certainly result in more language
prefixes than actually exist in the smaller projects. This is harmless though,
as requests for undefined prefixes will result in an error too.

Change-Id: I3e9c03566c43603590f369944db0707131435f66