Bugs 50603, 50589: Handle nested templates as well
This adds a slightly hacky work-around that enabled the TD fixup even for
nested transclusions as found on
http://en.wikipedia.org/wiki/5_O%27Clock_%28T-Pain_song%29.

We don't encapsulate nested transclusion content on the DOM, but (currently)
leave in data-parsoid.src. This patch uses that to detect the
|{{echo|{{!}}foo}} case.

In the longer term we'd like to remove data-parsoid.src for nested content. We
should think about a cleaner way to handle this, potentially by marking up
nested templates somehow.

Change-Id: Ib3cdaaf2712f7a7513af47df14a46a67d01fb407