Record performance numbers during RT testing
Add a perfstats table to keep different performance metrics for each
page and commit. They are kept as key, value pairs so that it'll be easy to
add new stats in the future. The stats are passed from the client to the
server embedded in the test results, using a new tag.

There's also a very basic interface to view the stats at /perfstats.

Bug: 46659
Change-Id: If73731d53cf360b6a25b3d7bb2bc5c30a2a4e4b2