Bug 53407: Don't modify the document after parsing
Our document would be modified right after the callback had fired, which made
further asynchronous processing harder than necessary. This patch delays the
cleanup until the next token is received.

With this in place we can properly sequence gzipping and roundtrip diffing in
tests so that page sizes are always recorded.

Change-Id: I14aa49a86c16219d3340c355aebf9b5b60e37952