Don't strip trailing newlines and space on tokens from sub-pipelines
The stripEOFTkfromTokens method used to strip trailing newlines and space-only
string tokens from sub-pipelines in addition to the expected EOFTk. This might
have been a work-around for missing paragraph suppression and possibly missing
stripping in other places, but does not seem to be needed any more.

It did create non-deterministic output when a trailing space-only token was
emitted from a sub-pipeline. The grouping of tokens returned from the
sub-pipeline depended on the order of template expansion, so the space would
only be stripped in some execution orders.

I have created a (currently unused) replacement method
stripTrailingNewlinesFromTokens that does the newline stripping in case that
still turns out to be necessary. We should remove it when we are satisfied
that the stripping is really not needed.

Change-Id: I83de5e87387d4d629eece946633e91b8e0f30854