Bug 53071: Handle about-less images better
* Only transfer about if there is a non-null about attribute in
  unpackDOMFragments

* Don't consume siblings if there is no about in getAboutSiblings

I am not 100% sure that this is indeed the root cause for bug 53071. A
scenario I can think of is this:

1) getAboutSiblings returned sibling nodes without about attributes
2) for some reason that is unclear to me those were *not* assigned
   about="null" attributes in unpackDOMFragments on reuse
3) the serializer faithfully serialized out the new content

The diffs in
https://it.wikipedia.org/w/index.php?title=Trieste&diff=prev&oldid=61046913
would support this, as the entire image link is duplicated.

The diff in https://it.wikipedia.org/w/index.php?title=Dalmine&oldid=61238456
however looks like an incomplete image duplication. There are several
figure-related DSR errors when parsing that page of this pattern:

WARNING: DSR inconsistency: cs/s mismatch for node: FIGURE s: 22704; cs: 22817
WARNING: DSR inconsistency: cs/s mismatch for node: FIGURE s: 18478; cs: 18586

Those warnings are still there when parsing the page with fragment reuse
enabled, so this might be a different bug.

Change-Id: I5a273aa39904230cae8500e5ada78c5e630e443d