Bug 53559: Fix problems with rt_testing results storing and entity escaping
There are two problems that contribute to the errors seen in Chrome for
some rt_testing page results:
- First, as reported on the bug, the client escapes all entities. We're using
UTF-8 throughout, so this is not necessary and takes up much more space than
needed.
- Second, TEXT fields in MySQL are capped at 64kb, which makes them too
small for some of the results. The result is clipped, producing the
bad-formed XML Chrome complains about.

To fix it, this patch adds code to only encode the strictly-necessary entities
and a SQL script to alter the results column to a MEDIUMTEXT field, which
should provide storage enough.

Bonus: Get the page's prefix correctly from the environment and remove
the total time timer from perfstats.

Change-Id: I23d02ee5c33fdf404a19ed00932d394f03040755