Don't emit foster box in transclusion
In d05894f170ba8e9a712099d3e6e6cd48ff53808d we removed encapsulation when
the associated table was in transclusion. This goes one step further and
avoids emitting it altogether. It shouldn't be needed and is breaking in
the odd unbalanced table situations.

Adds an assertion that the table is a sibling of the foster box.

Bug: 55820
Change-Id: I8424f4c3ec4de6c2b9072cb1712ae0ca4b312536