Disable persistent connections in curl client
This is a port of https://gerrit.wikimedia.org/r/#/c/92010/, which fixed a
performance regression in core. This might be specific to a broken HTTP
pipelining implementation in Swift, but for good measure lets port this to
our curl client too.

TODO:
Develop a shared curl_multi client in core

Change-Id: I92d3d4e20dc59b07b2e7964e5b979ca05ff71250