Bug 56282: Enable gzip compression support, the sequel
Varnish will

* always send 'Accept-Encoding: gzip' to the backend
* ignore a returned 'Vary: Accept-Encoding' header
* transparently decompressed stored gzipped content

Verified working using the varnish setup in betalabs.

The Parsoid version in betalabs is also old enough to still send
the cookie header unconditionally in requests back to varnish, and
thus exhibits the same cache misses we saw in the last deployment.
I manually tested with varying accept headers, but without the 
cookie header.

Change-Id: If94b65564ab7264969636559f9c79efc114ccfc1