server: make master and worker exit(0)
For some reason the master and workers process are exiting with error
code 1.  Seems to me that is the normal behavior and thus it deserves an
exit code of 0.

That would let us kill the server and keep proceeding in Jenkins (which
is usually set to abort whenever a command exit code is not 0).

Change-Id: I25177f911ed54a488579958d0a6236ef36aa2597