Prevent duplicate tests in round-trip testing
There were some cases of sending the same page to different clients if all
the clients couldn't finish a whole batch in the cutoff time. Removing
duplicate pages prevents this.

Also, clear the number of times a page has been tested when a new commit
is started if the page went over the maximum number of tries. This wasn't
done so crashers were marked as having being tested too many times.

Change-Id: I4622de7e6a073c529790b31bafb63fa64a9b489e