Add a new config option for an API proxy endpoint
Create a new configuration directive, to be passed through as the
"proxy" option to the requests module on calls that hit the API
endpoints.

This can be used to provide either a forward proxy (e.g. for some weird
isolated environment scenario) or, in Wikimedia's case, to provide the
service IP for the API application servers, as to be able to connect
directly to them, rather than going through the multiple caching layers
in front of them.

Hence, remove the "?random=<pid>" hacks that have been introduced with
Ie9554f8ea7f19b3271857bdc10e7ea6fa808dec to help with internal
load-balancing between caching layers, as they shouldn't be needed
anymore.

While at it, rename requests' "url" option to "uri", as this is the
documented and preferred way; the code internally maps "url" to "uri",
as it's a common mistake, but we shouldn't make it.

Bug: 51273
Change-Id: I76c7a133d05f432d692fb82df196c4a1051e2e43