Bug 51273: Support per-wiki API proxies
We had to disable direct access to the API cluster in production as that was
breaking private wikis and wikis not hosted on the same cluster (wikitech).

This patch adds the ability to configure API proxy settings per wiki. This
lets us disable proxying for those wikis. We could also proxy them through
another service.

Change-Id: I042b8b091652374696c76aafe3964eca80dfc662