Bug 58355: Accept html pre with attribute and trailing space
This reverts I6a63144f58f58f72c8df7a00db5589c09026a675, which results in the
wrong parse result. The html pre production was simply not accepting space
after attributes, and the previous patch then broke the inline production so
that the pre was allowed to parse as a generic inline tag instead.

Also adds a passing test.

Change-Id: Ia78fea860cf7fd4a46d84d16798b92569bf1cb9b