Handle fostered foster boxes
In certain situations, foster boxes end up being fostered themselves
and, to be doubly duplicitous, get nested so that they don't have a
table as a sibling.

In c5ee7f20969b37adec7e00e81444da5ae39d5317, we saw how transclusions
can affect the nesting around foster boxes and decided not to emit them
altogether in transclusion.

Here we switch from emitting meta tags to self-closed tables as the
foster-box marker tags. Since the HTML building algorithm never fosters
table tags, this new solution has the benefit of never fostering the
foster-box marker tags. It is also safe because foster boxes are only
ever inserted adjacent to another table.

This resolves the problem of placing a table in a fosterable position of
another table, as in:

  {{Top}}
  <div>
  {||}

Or, see enwiki:Wikipedia:WikiProject_Numismatics which is rife with more
examples.

Bug: 58355
Change-Id: Iee984759cd79c2d7536d473749c0a8446e58453c