Tolerate interwikis without $1 placeholder
Those would previously throw off the InterWikiMatcher. Without the $1, 
no matching group would be inserted in the regexp, which would then
throw off the index of other interwiki prefixes. That lead to wrong 
interwiki prefixes being used in the serializer.

This patch treats those broken interwikis as if they had a trailing $1,
both in wt2html and html2wt conversions.

Change-Id: Id61a2ed5e551f2811e2af266ea7961889411cbc6