Initial refactoring of WSP image support.
We banish the huge wall of var declarations and move the declarations closer
to the uses.  This helps us see the scope of some of the variables and
identify some unused variables.

We also rework the initial "identify the DOM structure" code to match the
One True Structure of images, and give the four parts better names.
This simplifies some crazy DOM manipulation and confines it to the
start of the function.

This version of the code should be functionally identical to the previous
version (bugs and all) although it should be a little bit more reliable
when picking out <figure> structure from arbitrary input.

Still left to do: break apart the actual option parsing and make it
clear where we are using data-parsoid as a crutch.

Change-Id: I30184edf9c55cce6b4449a3ac397823be745827e