Support JSON-valued parsoid parserTest option.
The parsoid option is now a JSON object by default.  For ease when writing
parser tests (and compatibility with existing tests), `parsoid=xxx,yyy`
entries in `!!options` are treated as shorthand for:
  parsoid={ "modes": [ "xxx", "yyy" ] }

A bare `parsoid` option is treated as if it were `parsoid={}`.
This results in no change in behavior, but it allows us to handle
parsoid-related options more consistently.

Change-Id: I0adee57c3c32e26efe706fc76cc3cc93ceabd342