Add new wt2wt mode which mutates DOM between wt2html and html2wt.
This test mode is used to ensure that we properly ignore data-parsoid
when VE edits the DOM.  There were a number of image-related bugs
of this type.  By mutating the DOM between the wt2html and html2wt
stages, we avoid adding a dependency on a certain data-parsoid
format to the parser tests file.

Change-Id: I6cb3b13c196cfb17fd8fdd32d0e288959b64db89