Bug 52618: Rescue stripped tds outside of table context
This implements basic recovery for stripped wikitext tds at the start of a
line, outside of a table. It is not a full fix, but at least gets quite a bit
closer.

To be able to distinguish html-syntax from wikitext-syntax end tags, I added
data-parsoid to the shadows and used tsr directly from there instead of
including it in the custom data-stag encoding.

TODO:
- Avoid <nowiki>fication on round-trip
- Avoid paragraph splitting by moving the paragraph wrapper to the DOM (ha!)
- Maybe move this to a separate pass?

Change-Id: I5d357781830c6505a096b75f587e1e077a16cfe0