Add a _rtselser-like selser mode
This mode appends a known comment to the body to force selser to serialize the
top-level blocks independently. The known comment is stripped from the output,
and the result is compared to the original input rather than the (possibly
broken) output of non-selser wt2wt.

Most of the failures exposed by this mode fall in the same class of errors
(failed expansion error message in native template expansion mode not
round-tripping), so should a) be easy to fix and b) not happen in production.

Other tests are expected to not round-trip (unclosed comment), while again
others look like real bugs that are also observed in production ("HTML-pre: 1.
embedded newlines").

TODO: Whoever cleans up this parserTests mess will be a hero!

Change-Id: Ic2146d35b226c048c27f0d7a5f9c8d00695b398a