Match PHP parser behavior for image `frame` option.
The PHP parser behavior is probably buggy; see
https://bugzilla.wikimedia.org/show_bug.cgi?id=62258

But with this patch we document the current behavior and match it.
If/when the PHP parser is fixed, parsoid should be adjusted to match.

Note that part of the fix is making our MockAPI behave the same as PHP,
which refuses to scale up an image to larger than its original size, and
does some somewhat eccentric math in order to scale by height.

We also need to tweak the WTS to account for the fact that width is
ignored for framed images.

Bug: 53514
Change-Id: I29d01e2732e02a9c356b645476f134c601e8d1dd