Reinstate disconnect
Some more testing showed that we are still starting the server using cluster,
which means that a normal process.exit will lead to a restart of the worker
with the *same* port. Instead we want the master to quit, so that apiServer
can re-try with another port.

The issue observed in rt testing must be caused by something else,
investigating that.

Change-Id: I45f579c60457f122829ae3496252e6d02d79772e