Add parsoid equivalents of additional image-related parser tests.
"SVG thumbnails with invalid language code" is a new test failure, because
parsoid doesn't validate the language code for the image `lang` option.

We change a few [[Image:foobar.jpg]] to [[File:Foobar.jpg]] to ensure
that the html2wt tests will work; this is a stronger test than just
disabling html2wt.  Where there are image options we still need to
disable html2wt, though, since the HTML representation does not
specify the order of the image options.

Change-Id: I3e202de6ab58faa56b356538c3b652c79dbecd97