No more NaNxNaN px.
There is a data-parsoid.a entry for `width` and `height` but no
data-parsoid.sa entry.  That causes `DU.getAttributeShadowInfo` to
return bogus information; fix it so that it doesn't assume that
a data-parsoid.sa entry must exist if the data-parsoid.a entry exists.

Clean up figureHandler to remove some unnecessary try/catch blocks.

Add tests for bug 62805 using the new `wikitext/edited` framework.

Bug: 62805
Change-Id: Ib85321b8a5adc6006a198f0d9678f852943fbc1c