Add parsoid outputs to "bogus format options" test case.
We also add a bit of support for recording these bogus options in
data-parsoid which lets us round-trip them.  This also allows us to
collect statistics on bogus option use.

A number of new selser tests pass because round-tripping is fixed.
We have some failures on the newly-added test case due to the well-known
issue with selser inserting content outside <p> containers.

Change-Id: I5743365c0c93fb8b256abd3dff758d6e18601bd3