Fix RT crasher caused by pipe characters in link text.
Added parser test to exercise this case, so we won't have to use RT to
catch this in the future.  This test case fails wt2wt, html2html, and
html2wt currently because it adds extra <nowiki> tags to the link content
during WTS (which manifest as <span> tags in the html).  These <nowiki>s
aren't strictly necessary, as discussed for the `BUG 2095` test case in
36053d3b568145d991d5088e1dbe561d60241f02.  So I'm leaving the failures
on the blacklist to indicate that we might fix this in the future.

Change-Id: If047537f106c0a7ae7b54ac69230345187b00ff3