Support protocol-relative URLs in the interwiki map.
A recent upstream change (https://gerrit.wikimedia.org/r/134264)
marked protocol-relative interwikimap entries in the siteinfo API
response.  Use this information to generate protocol-relative URLs for
interwiki links and language links, and to recognize either the
protocol-relative form, or the http/https forms of the URL in WTS.

In the process get rid of some old code which added language-based
pseudo-entries to the interwiki map.  This isn't done on the PHP
side and just confuses debugging, since our interwiki map won't match
the target wiki.  Tweak the parserTests framework to create the
necessary interwiki entries for parserTests the same way the PHP
parser test framework does.

Change-Id: I14783c1c257d23e60a177399b22bfbbefd48357c