Bug 57670: Avoid recursion via nested_block_in_table
So far we didn't have any protection against matching remaining table_lines
via the nested_block_in_table production. This lead to stack overflows and
slow parsing of large tables, as the remainder of the table would be parsed
recursively instead of iteratively.

This patch adds a syntactical stop and tests this for most table_lines
productions. This is fine as we don't care about nesting in the tokenizer. All
table lines are just parsed as flat tokens, to be assembled later by the
treebuilder.

Tested lightly with Obama etc, but will benefit from rt testing.

Change-Id: I0cf2302e209a5bfac8c81ff67ae160bf801a2fee