Better indent-pre / table interaction heuristic
489f16b5 removed greedy matching of consecutive indent-pre lines. This led to
indented table-like syntax like

 a
 | b

no longer being recognized as indented pre. This broke the rendering of pages
like eswiki/Herodes_I_el_Grande.

This patch reinstates the greedy indent-pre matching, but adds a protection
which causes the production to move back to block_lines if a table start tag
is found. There are also two new tests, one passing and one failing. The
failing case starts the indent-pre with table-like syntax:

a
 | b

This can't generally be handled in the tokenizer. It could be handled on the
token stream or the DOM.

Change-Id: I588526b9fbf7a5e3ea4a7299208f7057dcc3dabd