Revert "Use TokenStreamPatcher to convert table tags outside wikitables to text"
This reverts commit b272677b89dd9aab3f2e2533efa5996491109a3a.

This should only be run at the top-level (whether on the final DOM or 
the final token stream) and needs pipeline information about this. We 
are currently missing this info and causes regressions on pages like 
http://localhost:8000/_rt/jawiki/%E3%82%A4%E3%82%BF%E3%83%AA%E3%82%A2%E3%83%95%E3%82%A3%E3%82%AE%E3%83%A5%E3%82%A2%E3%82%B9%E3%82%B1%E3%83%BC%E3%83%88%E9%81%B8%E6%89%8B%E6%A8%A9

I have a fix for identifying a top level pipeline, but it is better to 
add that explicit flag to the pipeline rather than use a hack that I
have in my unsubmitted patch. Since we need to deploy today, let us 
revert this and fix this later.

Change-Id: If5024cdb7127f2286d0c8c5abc4659f97c3d5e9c