Handle local interwiki links.
As part of a larger effort to allow transclusions from commons (bug 4547),
the localinterwiki flag has been added to siteinfo
(https://gerrit.wikimedia.org/r/141276).  This allows parsoid to use a
less-hacky way of identifying local links (bug 66709), and also fix a
variety of local-link test cases introduced recently in core (bug 69909).

Note that [[en:Foo]] should render differently than [[Foo]] -- the link
text for the former is `en:Foo` while for the latter it is `Foo`.
This patch fixes that issue as well (bug 45209).

We've also fixed the behavior of [[local:]], which should render as
a link to the main page of the wiki (bug 64167).

Also removed the `protorel` property on most entries in the interwikimap
used by parser tests, in order to more closely match the setup in
PHP parser tests.  This required adding `http:` to a few test cases.
Made the `mi` prefix a language (which differs from PHP) in order
to better exercise both language and non-language localinterwiki
prefixes.

Also adds `title` attributes to interwiki links, which subsumes the
patch in Id555f1b6e08930d3777d2d988e6d6819590c5c0c.

Bug: 45209
Bug: 66709
Bug: 69909
Change-Id: Icf4073cab059246e7d35ccaa50c967a9fc888c9e